HelpButton: be backward compatible with PVE help mapping
pve-doc-generator used pveOnlineHelp as variable name while the
widget toolkit expects proxmoxOnlineHelp, allow both for now.

As its not quite clear when the switch to the widget toolkit happens
this is more flexible than changing the generator and adding a
versioned build dependency in pve-manager for it.
We normally do not bump pve-docs during releases either.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>
CC: Dominik Csapak <d.csapak@proxmox.com>