change allowed v6 prefix-length to 128
Currently the frontend validation only allows values 8-120 as prefix-lengths for
ipv6-cidr notations.

Following commit 70ea22506939f9687e1146c0f35247cfbe9225b8 in pve-common, this
patch changes that value to 128, which, while seldomly seen in regular setups
is technically correct.

Problem was reported by a user in the forum - they needed to set a length of
125, because their provider supplies only smaller networks. [0]

[0] https://forum.proxmox.com/threads/lxc-ipv6-why-8-120.52638/

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>