LogView: also reduce log syscall on scrolling
move the "return early, because a request is already pending" into
the doLoad function, so that all calls to it underly the request
limiting, but note that such a new request is desired, and if so
schedule it once the current finishes. This ensures that scrolling
will always get honored, but does not produces multiple hanging
workers, at max one.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>