task viewer: show task endtime and duration in status
But only when the caller gives us the endtime, since the status API
call does not give us the endtime.

As we stop updating the store once a task isn't running anymore the
fallback to current unix utc timestamp doesn't needs an additional
check, the last one is shown static then.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
[ Thomas: improve commit message ]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>