edit window: add extraParams config
For passing some default params which are passed along automatically
when calling load and doing the submit.

Useful in PBS, as there we often have the snapshot ID in multiple
parameters, not included in the request path, required for various
API call.

The params take low precendence, and are overwritten by values in
form fields with the same name for submit, or explicit params passed
to load calls

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>