css: improve icon rendering
Font Awesome is intended to be used at 14px font-size, we set that
already over product specific CSS for some elements like buttons.

But, for tab panel icons and tree navigation views it was often
overruled by ExtJS to 13px which made look them quite blurry and even
the position was a bit off (to high when compared to the texts
baseline.

Fix that by overwriting the font-size to 14px of the
x-treelist-item-icon and x-tab-icon-el-default class, for the latter
improve the icon to text possition by setting a slight top margin.

Setting the color to black was already done by each product, so move
that over here.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>