input panel: make advanced column check more flexible
allow that any of the three column/docked definitions can be set,
without setting the first one to an empty array to go in that if
branch.

Note, column2 works fine, but I did not greatly test a sole
advancedColumnB definition, so there may be still some improvements
for fixes - but at that point the user could also use advancedItems
to have full control.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>