dark-mode: set intentionally black icons to `$icon-color`
some icons intentionally use black as their color in the light theme.
this includes the little pencil and check mark icon in the acme
overview. change their color to the regular dark-mode icon-color. for
this to work the filter inversion needed for some other icons needs to
be removed too.

Signed-off-by: Stefan Sterz <s.sterz@proxmox.com>